Switch from hierarchal hashtables to two tree maps
The previous implementation of the environment consisted of a
mutable hash table and an optional parent environment. The new
environment implementation consists of two immutable tree maps,
one for the current frame and the other for all of the previous
frames.

The change caused major refactoring in desugar.ml. The environment
was previously mutable and a field of Desugar.t. Because the
environment is now immutable, its updates are passed around
explicitly as new values.

The rationale for this change is that the previous implementation
of the environment would have the performance of an associated
list, rather than a hash table, in a typical ML program,
characterized by many nested let expressions with each one only
introducing a few bindings. In the new implementation, all of the
parent environments' bindings are stored in a single tree map
rather than a chain of hash tables.